net: move batched output stream to net::
