+++ src/v/CMakeLists.txt
@@ -11,6 +11,7 @@ add_subdirectory(bytes)
 add_subdirectory(compression)
 add_subdirectory(utils)
 add_subdirectory(syschecks)
+add_subdirectory(net)
 add_subdirectory(rpc)
 add_subdirectory(model)
 add_subdirectory(json)
