+++ src/v/net/CMakeLists.txt
@@ -1,0 +1,7 @@
+v_cc_library(
+  NAME net
+  SRCS
+    batched_output_stream.cc
+  DEPS
+    Seastar::seastar
+  )
