+++ src/v/rpc/connection.h
@@ -11,7 +11,7 @@
 
 #pragma once
 
-#include "rpc/batched_output_stream.h"
+#include "net/batched_output_stream.h"
 #include "rpc/server_probe.h"
 #include "seastarx.h"
 
@@ -51,7 +51,7 @@ private:
     ss::sstring _name;
     ss::connected_socket _fd;
     ss::input_stream<char> _in;
-    batched_output_stream _out;
+    net::batched_output_stream _out;
     server_probe& _probe;
 };
 } // namespace rpc
